update getVault description
